+++ src/env.rs
@@ -393,9 +393,10 @@ pub fn get_cargo_workspace(manifest_dir: &str) -> Arc<PathBuf> {
             .current_dir(manifest_dir)
             .output()
             .unwrap();
-            let docs =
-                yaml_rust::YamlLoader::load_from_str(std::str::from_utf8(&output.stdout).unwrap())
-                    .unwrap();
+            let docs = crate::content::yaml::vendored::yaml::YamlLoader::load_from_str(
+                std::str::from_utf8(&output.stdout).unwrap(),
+            )
+            .unwrap();
             let manifest = docs.first().expect("Unable to parse cargo manifest");
             let workspace_root = PathBuf::from(manifest["workspace_root"].as_str().unwrap());
             Arc::new(workspace_root)